+++ python/tests/unit/eyes_selenium/test_nmg_capabilities.py
@@ -1,63 +1,108 @@
-from applitools.common import ProxySettings
+import pytest
+
+from applitools.common import EyesError, ProxySettings
 from applitools.selenium import Eyes
 
-API_KEY = "asd123"
-SERVER_URL = "https://eyesapi.applitools.com"
-PROXY_URL = "http://127.0.0.1:8888"
-PROXY_SETTINGS = ProxySettings(PROXY_URL)
-LIBRARY_PATH = "@executable_path/Frameworks/UFG_lib.xcframework/ios-arm64/UFG_lib.framework/UFG_lib:@executable_path/Frameworks/UFG_lib.xcframework/ios-arm64_x86_64-simulator/UFG_lib.framework/UFG_lib"
+LIBRARY_PATH = "@executable_path/Frameworks/UFG_lib.xcframework/ios-arm64_x86_64-simulator/UFG_lib.framework/UFG_lib:@executable_path/Frameworks/UFG_lib.xcframework/ios-arm64/UFG_lib.framework/UFG_lib"
 
 
-def test_android_nmg_capabilities():
-    caps = {}
-    eyes = Eyes()
-    eyes.set_nmg_capabilities(caps, API_KEY, SERVER_URL, PROXY_SETTINGS)
+def test_nmg_capabilities_explicit():
+    caps = {"capsKey": "capsValue"}
+    Eyes.set_nmg_capabilities(caps, "abc", "https://server", "http://proxy:1234")
 
-    android_args = caps["optionalIntentArguments"]
-    assert (
-        android_args
-        == """--es APPLITOOLS '{"NML_API_KEY": "%s", "NML_PROXY_URL": "%s", "NML_SERVER_URL": "%s"}'"""
-        % (
-            API_KEY,
-            str(PROXY_SETTINGS.url),
-            SERVER_URL,
-        )
-    )
-
-
-def test_ios_nmg_capabilities():
-    caps = {}
-    eyes = Eyes()
-    eyes.set_nmg_capabilities(caps, API_KEY, SERVER_URL, PROXY_SETTINGS)
-
-    ios_args = caps["processArguments"]
-    assert ios_args
-    assert ios_args["args"] == []
-
-    ios_envs = ios_args["env"]
-    assert ios_envs["NML_API_KEY"] == API_KEY
-    assert ios_envs["NML_SERVER_URL"] == SERVER_URL
-    assert ios_envs["NML_PROXY_URL"] == PROXY_SETTINGS.url
-    assert (
-        ios_envs["DYLD_INSERT_LIBRARIES"]
-        == "@executable_path/Frameworks/UFG_lib.xcframework/ios-arm64_x86_64-simulator/UFG_lib.framework/UFG_lib:@executable_path/Frameworks/UFG_lib.xcframework/ios-arm64/UFG_lib.framework/UFG_lib"
-    )
+    assert caps == {
+        "capsKey": "capsValue",
+        "optionalIntentArguments": '--es APPLITOOLS \'{"NML_API_KEY": "abc", '
+        '"NML_PROXY_URL": "http://proxy:1234", '
+        '"NML_SERVER_URL": "https://server"}\'',
+        "processArguments": {
+            "args": [],
+            "env": {
+                "DYLD_INSERT_LIBRARIES": LIBRARY_PATH,
+                "NML_API_KEY": "abc",
+                "NML_PROXY_URL": "http://proxy:1234",
+                "NML_SERVER_URL": "https://server",
+            },
+        },
+    }
 
 
-def test_defaults_from_env(monkeypatch):
-    monkeypatch.setenv("APPLITOOLS_API_KEY", API_KEY)
-    monkeypatch.setenv("APPLITOOLS_SERVER_URL", SERVER_URL)
-    monkeypatch.setenv("APPLITOOLS_HTTP_PROXY", PROXY_URL)
+def test_nmg_capabilities_explicit_proxy_settings():
+    caps = {"capsKey": "capsValue"}
+    Eyes.set_nmg_capabilities(caps, "abc", proxy_settings=ProxySettings("host", 81))
 
-    caps = {}
-    eyes = Eyes()
-    eyes.set_nmg_capabilities(caps)
+    assert caps == {
+        "capsKey": "capsValue",
+        "optionalIntentArguments": '--es APPLITOOLS \'{"NML_API_KEY": "abc", '
+        '"NML_PROXY_URL": "http://host:81"}\'',
+        "processArguments": {
+            "args": [],
+            "env": {
+                "DYLD_INSERT_LIBRARIES": LIBRARY_PATH,
+                "NML_API_KEY": "abc",
+                "NML_PROXY_URL": "http://host:81",
+            },
+        },
+    }
 
-    ios_args = caps["processArguments"]
-    assert ios_args
-    assert ios_args["args"] == []
 
-    ios_envs = ios_args["env"]
-    assert ios_envs["NML_API_KEY"] == API_KEY
-    assert ios_envs["NML_SERVER_URL"] == SERVER_URL
-    assert ios_envs["NML_PROXY_URL"] == PROXY_SETTINGS.url
+def test_nmg_capabilities_explicit_api_key_only(monkeypatch):
+    monkeypatch.delenv("APPLITOOLS_SERVER_URL", False)
+    monkeypatch.delenv("APPLITOOLS_HTTP_PROXY", False)
+    caps = {"capsKey": "capsValue"}
+    Eyes.set_nmg_capabilities(caps, "abc")
+
+    assert caps == {
+        "capsKey": "capsValue",
+        "optionalIntentArguments": '--es APPLITOOLS \'{"NML_API_KEY": "abc"}\'',
+        "processArguments": {
+            "args": [],
+            "env": {"DYLD_INSERT_LIBRARIES": LIBRARY_PATH, "NML_API_KEY": "abc"},
+        },
+    }
+
+
+def test_nmg_capabilities_from_env(monkeypatch):
+    monkeypatch.setenv("APPLITOOLS_API_KEY", "def")
+    monkeypatch.setenv("APPLITOOLS_SERVER_URL", "https://otherserver")
+    monkeypatch.setenv("APPLITOOLS_HTTP_PROXY", "http://secondproxy:2222")
+    caps = {"capsKey": "capsValue"}
+    Eyes.set_nmg_capabilities(caps)
+
+    assert caps == {
+        "capsKey": "capsValue",
+        "optionalIntentArguments": '--es APPLITOOLS \'{"NML_API_KEY": "def", '
+        '"NML_PROXY_URL": "http://secondproxy:2222", '
+        '"NML_SERVER_URL": "https://otherserver"}\'',
+        "processArguments": {
+            "args": [],
+            "env": {
+                "DYLD_INSERT_LIBRARIES": LIBRARY_PATH,
+                "NML_API_KEY": "def",
+                "NML_PROXY_URL": "http://secondproxy:2222",
+                "NML_SERVER_URL": "https://otherserver",
+            },
+        },
+    }
+
+
+def test_nmg_capabilities_from_env_api_key_only(monkeypatch):
+    monkeypatch.setenv("APPLITOOLS_API_KEY", "def")
+    caps = {"capsKey": "capsValue"}
+    Eyes.set_nmg_capabilities(caps)
+
+    assert caps == {
+        "capsKey": "capsValue",
+        "optionalIntentArguments": '--es APPLITOOLS \'{"NML_API_KEY": "def"}\'',
+        "processArguments": {
+            "args": [],
+            "env": {"DYLD_INSERT_LIBRARIES": LIBRARY_PATH, "NML_API_KEY": "def"},
+        },
+    }
+
+
+def test_nmg_capabilities_no_api_key_raises_exception(monkeypatch):
+    monkeypatch.delenv("APPLITOOLS_API_KEY", False)
+    caps = {"capsKey": "capsValue"}
+    with pytest.raises(EyesError):
+        Eyes.set_nmg_capabilities(caps)