+++ python/eyes_robotframework/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Fixed
+- IosVersion.LATEST is not a valid AndroidVersion error when native_mobile_grid is configured with Android device [Trello 3174](https://trello.com/c/O3c1xs0L)
+
 ## [5.11.0] - 2022-12-21
 ### Added
 - Configuration option dont_close_batches can now be set in yaml config [Trello 3103](https://trello.com/c/kxOBjjCs)